Update `PackUri` and remove `AssemblyResourceUtils` implementation.
