+++ Directory.Build.props
@@ -1,5 +1,5 @@
 <Project>
   <PropertyGroup>
-    <Version>0.4.1</Version>
+    <Version>0.5.0-alpha</Version>
   </PropertyGroup>
 </Project>
